+++ saleor/checkout/complete_checkout.py
@@ -23,7 +23,7 @@ from ..discount.utils import (
     increase_voucher_usage,
     remove_voucher_usage_by_customer,
 )
-from ..order import OrderStatus
+from ..order import OrderLineData, OrderStatus
 from ..order.actions import order_created
 from ..order.models import Order, OrderLine
 from ..order.notifications import send_order_confirmation, send_staff_order_confirmation
@@ -32,7 +32,7 @@ from ..payment.models import Payment, Transaction
 from ..payment.utils import store_customer_id
 from ..product.models import ProductTranslation, ProductVariantTranslation
 from ..warehouse.availability import check_stock_quantity_bulk
-from ..warehouse.management import allocate_stock
+from ..warehouse.management import allocate_stocks
 from . import AddressType, models
 from .checkout_cleaner import clean_checkout_payment, clean_checkout_shipping
 from .models import Checkout
@@ -136,7 +136,7 @@ def _create_line_for_order(
     channel: "Channel",
     products_translation: Dict[int, Optional[str]],
     variants_translation: Dict[int, Optional[str]],
-) -> OrderLine:
+) -> OrderLineData:
     """Create a line for the given order.
 
     :raises InsufficientStock: when there is not enough items in stock for this variant.
@@ -195,7 +195,9 @@ def _create_line_for_order(
         tax_rate=tax_rate,
     )
 
-    return line
+    line_info = OrderLineData(line=line, quantity=quantity, variant=variant)
+
+    return line_info
 
 
 def _create_lines_for_order(
@@ -204,7 +206,7 @@ def _create_lines_for_order(
     lines: Iterable["CheckoutLineInfo"],
     discounts: Iterable[DiscountInfo],
     channel: "Channel",
-) -> Iterable[OrderLine]:
+) -> Iterable[OrderLineData]:
     """Create a lines for the given order.
 
     :raises InsufficientStock: when there is not enough items in stock for this variant.
@@ -329,6 +331,7 @@ def _create_order(
     order_data: dict,
     user: User,
     plugin_manager: "PluginsManager",
+    site_settings=None
 ) -> Order:
     """Create an order from the checkout.
 
@@ -348,10 +351,11 @@ def _create_order(
         return order
 
     total_price_left = order_data.pop("total_price_left")
-    order_lines = order_data.pop("lines")
+    order_lines_info = order_data.pop("lines")
 
-    # TODO: refactor to use request.site / info.context site
-    site_settings = Site.objects.get_current().settings
+    if site_settings is None:
+        site_settings = Site.objects.get_current().settings
+
     status = (
         OrderStatus.UNFULFILLED
         if site_settings.automatically_confirm_all_new_orders
@@ -363,15 +367,16 @@ def _create_order(
         status=status,
         channel=checkout.channel,
     )
-    for line in order_lines:
+    order_lines = []
+    for line_info in order_lines_info:
+        line = line_info.line
         line.order_id = order.pk
-    order_lines = OrderLine.objects.bulk_create(order_lines)
+        order_lines.append(line)
 
-    # allocate stocks from the lines
-    for line in order_lines:  # type: OrderLine
-        variant = line.variant
-        if variant and variant.track_inventory:
-            allocate_stock(line, checkout.get_country(), line.quantity)
+    OrderLine.objects.bulk_create(order_lines)
+
+    country_code = checkout.get_country()
+    allocate_stocks(order_lines_info, country_code)
 
     # Add gift cards to the order
     for gift_card in checkout.gift_cards.select_for_update():
@@ -473,7 +478,9 @@ def _get_order_data(
             discounts=discounts,
         )
     except InsufficientStock as e:
-        raise ValidationError(f"Insufficient product stock: {e.item}", code=e.code)
+        raise ValidationError(
+            f"Insufficient product stock: {', '.join(e.items)}", code=e.code
+        )
     except NotApplicable:
         raise ValidationError(
             "Voucher not applicable",
@@ -523,6 +530,7 @@ def complete_checkout(
     store_source,
     discounts,
     user,
+    site_settings=None,
     tracking_code=None,
     redirect_url=None,
 ) -> Tuple[Optional[Order], bool, dict]:
@@ -571,12 +579,15 @@ def complete_checkout(
                 order_data=order_data,
                 user=user,  # type: ignore
                 plugin_manager=manager,
+                site_settings=site_settings,
             )
             # remove checkout after order is successfully created
             checkout.delete()
         except InsufficientStock as e:
             release_voucher_usage(order_data)
             gateway.payment_refund_or_void(payment)
-            raise ValidationError(f"Insufficient product stock: {e.item}", code=e.code)
+            raise ValidationError(
+                f"Insufficient product stock: {', '.join(e.items)}", code=e.code
+            )
 
     return order, action_required, action_data